move Stateless Worklet related logic to a different module
